requirements.txt
appdirs==1.4.4
arrow==1.0.3
asgiref==3.3.1
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==5.5
distlib==0.3.1
Django==3.1.7
django-audit-fields==0.3.2
django-defender==0.8.0
django-environ==0.4.5
django-ipware==2.1.0
django-logentry-admin==1.0.6
django-redis==4.12.1
django-revision==0.1.20
django-simple-history==2.12.0
django-storages==1.11.1
django-webtest==1.9.7
docopt==0.6.2
edc-utils==0.3.1
Faker==7.0.1
filelock==3.0.12
flake8==3.9.0
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
isort==5.7.0
mccabe==0.6.1
model-bakery==1.2.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
pipreqs==0.4.10
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
qrcode==6.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
six==1.15.0
smmap==4.0.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tox==3.23.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
waitress==2.0.0
WebOb==1.8.7
WebTest==2.0.35
wifi-password==1.0.7
yarg==0.1.9